pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sfce"
version = "2.0.0"
description = "Sistema de Facturacion y Contabilidad Evolutivo"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
    "mistralai>=1.0",
    "openai>=1.0",
    "google-genai>=1.0",
]

[project.optional-dependencies]
db = ["sqlalchemy>=2.0", "alembic>=1.13"]
api = ["fastapi>=0.110", "uvicorn>=0.27", "websockets>=12.0", "pyjwt>=2.8"]
dev = ["pytest>=8.0", "pytest-cov>=4.1"]

[tool.setuptools.packages.find]
include = ["sfce*"]
